+++ src/xr_3da/xrGame/Torch.cpp
@@ -260,7 +260,7 @@ BOOL CTorch::net_Spawn(CSE_Abstract* DC)
 	
 	SwitchNightVision		(false);
 
-	m_delta_h				= PI_DIV_2- (float) atan((range*0.5f)/_abs(TORCH_OFFSET.x));
+	m_delta_h				= PI_DIV_2-atan((range*0.5f)/_abs(TORCH_OFFSET.x));
 
 	return					(TRUE);
 }
