+++ src/xr_3da/xrGame/ik/eqn.h
@@ -79,7 +79,7 @@ struct PsiEquation
     { 
 		alpha = a; beta = b; xi = x;  
 		a2b2 = a*a + b*b;
-		atan2ba = (float) atan2(b,a);
+		atan2ba = atan2(b,a);
 		num_crits = num_roots = status = 0;
 		status_ptr = &status;
 		num_roots_ptr = &num_roots;
