+++ src/xr_3da/xrGame/ik/jtlimits.cxx
@@ -391,7 +391,7 @@ inline float mytan(float v)
 	else
 	    v = 3*M_PI/2 - big_eps;
     }
-    return (float) tan(v);
+    return tan(v);
 }
 
 
@@ -435,12 +435,12 @@ static AngleInt PosCos(-M_PI/2,M_PI/2);
  
 float ComplexJtLimit::theta1(float v) const
 {
-    return angle_normalize((float) atan2(sin_eq.eval(v), cos_eq.eval(v)));
+    return angle_normalize(atan2(sin_eq.eval(v), cos_eq.eval(v)));
 }
 
 float ComplexJtLimit::theta2(float v) const
 {
-    return angle_normalize((float) atan2(-sin_eq.eval(v), -cos_eq.eval(v)));
+    return angle_normalize(atan2(-sin_eq.eval(v), -cos_eq.eval(v)));
 }
 
 float ComplexJtLimit::theta(int family, float psi) const
@@ -733,7 +733,7 @@ void ComplexJtLimit::store_intersections(int n,
     sort(n2, f2);
 }
 
-const float tan_0 = (float) tan(eps/2.0f);
+const float tan_0 = tan(eps/2.0f);
 const float tan_two_pi = (2*M_PI-eps/2.0f);
